% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_Lighter.R
\name{draw_Lighter}
\alias{draw_Lighter}
\title{Draw Lighter relay states}
\usage{
draw_Lighter(filename)
}
\arguments{
\item{filename}{A character string giving path and filename where input data has been stored.}
}
\value{
A ggplot2 object
}
\description{
This is a function that draw Ligther relays states dutring the day.
}
